Move SumUDF to UDF folder and improve documentation
