+++ docs/source/changelog/misc/sum.rst
@@ -1,0 +1,4 @@
+[Misc] Move SumUDF to UDF folder
+================================
+
+* :class:`~libertem.udf.sum.SumUDF` moved to the :mod:`libertem.udf` folder (:pr:`613`).
