+++ docs/source/reference/udf.rst
@@ -3,6 +3,8 @@
 UDF API reference
 -----------------
 
+.. _`define udf ref`:
+
 Defining UDFs
 ~~~~~~~~~~~~~
 
@@ -13,6 +15,8 @@ See :ref:`user-defined functions` for an introduction and in-depth explanation.
    :special-members: __init__
    :exclude-members: UDFTask,UDFRunner
 
+.. _`run udf ref`:
+
 Running UDFs
 ~~~~~~~~~~~~
 
@@ -21,6 +25,8 @@ Two methods of :class:`libertem.api.Context` are relevant for running user-defin
 .. autoclass:: libertem.api.Context
    :members: run_udf,map
    :noindex:
+
+.. _`buffer udf ref`:
 
 Buffers
 ~~~~~~~
@@ -42,21 +48,37 @@ Some generally useful UDFs are included with LiberTEM:
 .. note::
     See :ref:`application api` for application-specific UDFs and analyses.
 
+.. _`sum udf`:
+
+.. autoclass:: libertem.udf.sum.SumUDF
+    :members:
+    :special-members: __init__
+
+.. _`logsum udf`:
+
 .. autoclass:: libertem.udf.logsum.LogsumUDF
     :members:
     :special-members: __init__
+
+.. _`stddev udf`:
 
 .. autoclass:: libertem.udf.stddev.StdDevUDF
     :members:
     :special-members: __init__
 
+.. _`sumsig udf`:
+
 .. autoclass:: libertem.udf.sumsigudf.SumSigUDF
     :members:
     :special-members: __init__
 
+.. _`masks udf`:
+
 .. autoclass:: libertem.udf.masks.ApplyMasksUDF
     :members:
     :special-members: __init__
+
+.. _`pick udf`:
 
 .. autoclass:: libertem.udf.raw.PickUDF
     :members:
